+++ frontend/src/components/vue-default/AppFooter.vue
