--- frontend/src/layouts/README.md
@@ -1,5 +1,0 @@
-# Layouts
-
-Layouts are reusable components that wrap around pages. They are used to provide a consistent look and feel across multiple pages.
-
-Full documentation for this feature can be found in the Official [vite-plugin-vue-layouts-next](https://github.com/loicduong/vite-plugin-vue-layouts-next) repository.
