+++ frontend/src/layouts/default.vue
@@ -3,9 +3,9 @@
     <router-view />
   </v-main>
 
-  <AppFooter />
+  <!-- <AppFooter /> -->
 </template>
 
 <script lang="ts" setup>
-  //
+//
 </script>
