--- frontend/src/pages/README.md
@@ -1,5 +1,0 @@
-# Pages
-
-Vue components created in this folder will automatically be converted to navigatable routes.
-
-Full documentation for this feature can be found in the Official [unplugin-vue-router](https://github.com/posva/unplugin-vue-router) repository.
