--- frontend/src/pages/me.vue
@@ -1,8 +1,0 @@
-<template>
-    <div><h1>This is my first page.</h1></div>   
-     <div><h1>This is my changes.</h1></div>
-     <div><h1>This is my second changes.</h1></div>
-          <div><h1>This is my third changes.</h1></div>
-
-
-</template>
